+++ Source/MotionType.swift
@@ -6,7 +6,17 @@
 //  Copyright © 2019 Vladislav Kondrashkov. All rights reserved.
 //
 
+/**
+ Defines direction for motion animation
+ */
 public enum MotionType {
+    /**
+     Indicates that motion animation would change position's Y coordinate
+     */
     case vertical
+
+    /**
+     Indicates that motion animation would change position's X coordinate
+     */
     case horizontal
 }