+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# Changelog for Lotus 0.2.0
+## Added
+* Documentation on [Github Pages](lotus-docs.github.io)
+
+## Changed
+* All easings became math functions
+* Custom easing use math function instead of CAMediaTimingFunction
+
+## Fixed
+* [LI-37 Type issues](https://github.com/lotus-ios/lotus/issues/37)
+
 # Changelog for Lotus 0.1.0
 ## Added
 * Tests